Dockerfile: Remove high UID/GID
When running Docker with user namespace remapping, the UID/GID of the
files must be in the 0 - 65535 range. One toolchain's tar file included
files outside of that range and thus I could not run the image on my
GitLab runners. This patch fixes that and the U-Boot CI completes just
fine in my environment.

Signed-off-by: Harm Berntsen <harm.berntsen@nedap.com>
CC: Tom Rini <trini@konsulko.com>